refactor and add copilor suggestions
