+++ flows/hades/strategus_plugin/flowutils.py
@@ -1,11 +1,14 @@
 # TODO: refactor flowUtils.py - that shares duplicate code from dataflow-ui/flowUtils.py
 # Solution: create a common utils for ui-flows as `ui-flow-utils`
 
+import os
+import re
 import importlib
 import numpy as np
 import pandas as pd
 
 from prefect.variables import Variable
+from prefect.artifacts import create_markdown_artifact
 from rpy2 import robjects as ro
 from rpy2.robjects import pandas2ri
 from rpy2.robjects.packages import importr
@@ -125,3 +128,41 @@ def install_r_packages_from_lockfile(lockfile_path):
             logger.info("R packages installed successfully from lockfile.")
         except Exception as e:
             logger.error(f"Error in installing R packages from lockfile: {str(e)}")
+
+
+def is_strategus_execution_successful(logFilePath: str) -> tuple[bool, str]:
+    logger = Logger()
+    logger.info('Checking strategus execution log for errors...')
+
+    with open(logFilePath, "r") as f:
+        lines = f.readlines()
+    summary_idx = None
+    for idx, line in enumerate(lines):
+        pattern = re.compile(r"─* execution summary ─*", re.IGNORECASE)
+        if pattern.search(line.lower()):
+            print("Found execution summary line:", line)
+            summary_idx = idx
+            break
+    msg = "Strategus execution failed, check log for details."
+    if summary_idx is None:
+        return False, msg
+    error_found = False
+    errorMsg = ""
+    pattern = re.compile(r"✖ .*Module", re.UNICODE)
+    for line in lines[summary_idx+1:]:
+        if pattern.search(line):
+            error_found = True
+            errorMsg += line.strip() + "\n"
+    return (not error_found, msg if error_found else "")
+
+def save_strategus_log_file(log_file_path: str):
+    logger = Logger()
+    if os.path.exists(log_file_path):
+        with open(log_file_path, "r") as f:
+            file_contents = f.read()
+            create_markdown_artifact(
+                key="strategus-analysis-logs",
+                markdown=file_contents
+            )
+    else:
+        logger.warning('Strategus log file not found and no logs were saved')
